+++ src/api/v1/services/restaurant.service.ts
@@ -37,6 +37,9 @@ const addRestaurant = async (restaurantData: Partial<IRestaurant>) => {
 
 const updateRestaurant = async (restaurantId: string, updateData: Partial<IRestaurant>) => {
     try {
+        if (updateData.chef) {
+            await Chef.findByIdAndUpdate(updateData.chef, { $push: { restaurants: restaurantId } })
+        }
         return await Restaurant.findByIdAndUpdate(restaurantId, updateData, { new: true, runValidators: true })
     } catch (err) {
         console.log('restaurant service => error updating a restaurant')
@@ -46,9 +49,16 @@ const updateRestaurant = async (restaurantId: string, updateData: Partial<IResta
 
 const removeRestaurant = async (restaurantId: string) => {
     try {
+        const restaurant = await Restaurant.findById(restaurantId)
+        if (!restaurant) {
+            throw new Error('Restaurant not found')
+        }
+        if (restaurant.chef) {
+            await Chef.findByIdAndUpdate(restaurant.chef, { $pull: { restaurants: restaurantId } })
+        }
         return await Restaurant.findByIdAndDelete(restaurantId)
     } catch (err) {
-        console.log('restaurant service => error removing a restaurant')
+        console.log('restaurant service => error removing a restaurant', err)
         throw err
     }
 }
